`Update methods and add new features in MyIngestionTelemetry.cs, Generator.cs, and RoslynExtensions.cs`
Updated `CallApi` method in `MyIngestionTelemetry.cs` to a boolean method and added a new partial method `OnPollingSucceeded`. Added `GenerateOperations` and `GenerateEvents` methods in `Generator.cs` with modifications to work with `IFieldSymbol` and check for partial definitions. Updated `MatchAttribute` method in `RoslynExtensions.cs` to work with `IMethodSymbol` and added a new method `MatchMethodAttribute`. Renamed and updated `GenerateEvent` method in `Generator.cs` to generate partial methods for events.