Updated project reference in TelemetrySrcGen.Playground.csproj
Modified the project reference to `TelemetrySrcGen.Abstractions.csproj` in the `TelemetrySrcGen.Playground.csproj` file. An `OutputItemType` attribute with the value "Analyzer" has been added, indicating that the output of the referenced project will now be treated as an analyzer by the referencing project.

thanks to https://stackoverflow.com/a/71397443/101260 for the solution to this bug!